Extract kodein modules to their own files
