+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
   <uses-permission android:name="android.permission.INTERNET"/>
 
   <application
-      android:name=".StarArchWarsApplication"
+      android:name=".infrastructure.StarArchWarsApplication"
       android:allowBackup="true"
       android:icon="@mipmap/ic_launcher"
       android:label="@string/app_name"
