+++ app/src/main/java/com/smassive/stararchwars/films/view/FilmsActivity.kt
@@ -1,42 +1,24 @@
 package com.smassive.stararchwars.films.view
 
-import android.arch.lifecycle.ViewModelProviders
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
 import android.view.View
-import com.github.salomonbrys.kodein.Kodein
-import com.github.salomonbrys.kodein.android.androidActivityScope
-import com.github.salomonbrys.kodein.bind
 import com.github.salomonbrys.kodein.instance
-import com.github.salomonbrys.kodein.scopedSingleton
-import com.github.salomonbrys.kodein.singleton
 import com.github.salomonbrys.kodein.with
 import com.smassive.stararchwars.R
 import com.smassive.stararchwars.base.extensions.enterListAnimation
 import com.smassive.stararchwars.base.view.BaseActivity
 import com.smassive.stararchwars.data.base.extensions.observeNonNull
 import com.smassive.stararchwars.films.viewmodel.FilmsViewModel
-import com.smassive.stararchwars.films.viewmodel.FilmsViewModelFactory
+import com.smassive.stararchwars.infrastructure.di.films.filmsActivityModule
 import kotlinx.android.synthetic.main.activity_films.toolbar
 import kotlinx.android.synthetic.main.content_films.filmsList
 import kotlinx.android.synthetic.main.content_films.filmsLoading
 
 class FilmsActivity : BaseActivity() {
 
-  override val activityModule = Kodein.Module {
-    bind<FilmsAdapter>() with scopedSingleton(androidActivityScope) {
-      FilmsAdapter()
-    }
-
-    bind<FilmsViewModelFactory>() with singleton {
-      FilmsViewModelFactory(instance(), instance())
-    }
-
-    bind<FilmsViewModel>() with scopedSingleton(androidActivityScope) {
-      ViewModelProviders.of(this@FilmsActivity, instance<FilmsViewModelFactory>()).get(FilmsViewModel::class.java)
-    }
-  }
+  override val activityModule = filmsActivityModule(this)
 
   private val filmsAdapter: FilmsAdapter by with(this as AppCompatActivity).instance()
   private val filmsViewModel: FilmsViewModel by with(this as AppCompatActivity).instance()
